moved fixtures mod to a file
